+++ fdb-extensions/src/main/java/com/apple/foundationdb/async/rtree/AbstractNode.java
@@ -1,5 +1,5 @@
 /*
- * Node.java
+ * AbstractNode.java
  *
  * This source file is part of the FoundationDB open source project
  *